requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.5.0
langdetect==1.0.9
